add interrupt based pwm of pwm
the chained timer approach was prone to leave the OC pin of the
slave timer in "1" state when the slave timer was disabled by
the master timer. Gated mode *plus* reset on switch-off does not
seem to exist, unfortunately. So we use the former master timer
to trigger an interrupt routine that will switch on/off the OC
pin of the former slave timer (carrier generation).